Fixed compatibility with Spring Boot 2.0
