+++ src/test/java/devbury/threadscope/integration/ServletRequestListenerTest.java
@@ -16,50 +16,33 @@
 
 package devbury.threadscope.integration;
 
-import devbury.threadscope.ThreadScopeManager;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.test.SpringApplicationConfiguration;
-import org.springframework.boot.test.TestRestTemplate;
-import org.springframework.boot.test.WebIntegrationTest;
-import org.springframework.context.annotation.Bean;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-
-import javax.servlet.ServletRequestEvent;
 
 import static org.junit.Assert.assertTrue;
 
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = {ServletRequestListenerTest.class})
-@WebIntegrationTest("server.port:0")
+@SpringBootTest(classes = {ServletRequestListenerTest.class, ThreadScopeManagerProvider.class}, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @EnableAutoConfiguration
 @ComponentScan
 public class ServletRequestListenerTest {
 
     static boolean requestDestroyedCalled = false;
 
-    @Value("${local.server.port}")
-    int port;
+    @Autowired
+    private TestRestTemplate testRestTemplate;
 
     @Test
     public void verifyServletRequestListenerRegisteredAndCalled() throws Exception {
-        TestRestTemplate restTemplate = new TestRestTemplate();
-        restTemplate.getForObject("http://localhost:{port}/test", String.class, port);
+        testRestTemplate.getForObject("/test", String.class);
         assertTrue(requestDestroyedCalled);
     }
 
-    @Bean
-    public ThreadScopeManager threadScopeManager() {
-        return new ThreadScopeManager() {
-            @Override
-            public void requestDestroyed(ServletRequestEvent servletRequestEvent) {
-                requestDestroyedCalled = true;
-                super.requestDestroyed(servletRequestEvent);
-            }
-        };
-    }
 }
